Moved base component to a different dir
